continuing to separate out components
