+++ client/app/components/Form/index.jsx
@@ -1,12 +1,7 @@
 // @flow
 import React from 'react';
-import { REQUIRES_DEFAULT } from '../Input/InputDefault';
 import { displayQuickCreate } from './quickCreate';
-import { TYPES as INPUT_TYPES } from '../Input/utils';
-import { FormInputs } from './FormInputs';
-import { Utils } from '../../utils';
 import css from './Form.scss';
-
 
 type Errors = { [string]: boolean } | {};
 
@@ -35,42 +30,21 @@ export class Form extends React.Component<Props, State> {
     this.myRefs = {};
   }
 
-
-handleError = (id: string, error: boolean) => {
-  const { errors } = this.state;
-  const newErrors = { ...errors };
-  newErrors[id] = error;
-  this.setState({ errors: newErrors });
-};
-
-  isInputError = (input: any) => {
-    const validType = REQUIRES_DEFAULT.includes(input.type) || input.type === 'textarea';
-    return (
-      validType
-      && input.required
-      && this.myRefs[input.id]
-      && !this.myRefs[input.id].value
-    );
+  handleError = (id: string, error: boolean) => {
+    const { errors } = this.state;
+    const newErrors = { ...errors };
+    newErrors[id] = error;
+    this.setState({ errors: newErrors });
   };
 
   onSubmit = (e: SyntheticEvent<HTMLInputElement>) => {
     // Get errors from inputs that were never focused
-    const { inputs, errors } = this.state;
+    const { errors } = this.state;
     const newErrors = { ...errors };
-    const newInputs = inputs.map((input: any) => {
-      const newInput = { ...input };
-      if (this.isInputError(newInput)) {
-        newInput.error = true;
-        newInput.value = this.myRefs[input.id].value;
-        newInput.myKey = Utils.randomString(); // Triggers state change in child component
-        newErrors[newInput.id] = true;
-      }
-      return newInput;
-    });
     const { noFormTagSubmit } = this.props;
     if (hasErrors(newErrors) > 0) {
       e.preventDefault();
-      this.setState({ inputs: newInputs, errors: newErrors });
+      this.setState({ errors: newErrors });
     } else if (noFormTagSubmit) {
       noFormTagSubmit();
     }
@@ -81,7 +55,15 @@ handleError = (id: string, error: boolean) => {
     this.onSubmit(e);
   };
 
-
+  displayInputs = (): any => {
+    const { inputs } = this.state;
+    return inputs.map((input: any) => {
+      if (input.type === 'quickCreate') {
+        return displayQuickCreate(input);
+      }
+      return null;
+    });
+  };
 
   render() {
     const { action, noFormTag, noFormTagRef } = this.props;
@@ -89,7 +71,7 @@ handleError = (id: string, error: boolean) => {
       // This will go into Dynamic Form
       return (
         <div className={css.form} ref={noFormTagRef}>
-         <FormInputs />
+          {this.displayInputs()}
         </div>
       );
     }
@@ -102,7 +84,7 @@ handleError = (id: string, error: boolean) => {
         method="post"
         action={action}
       >
-       <FormInputs />
+        {this.displayInputs()}
       </form>
     );
   }
